Перенес onPictureElementClick в big-picture.js
